\name{gets}
\alias{gets}
\alias{gets.default}

\title{General-to-Specific (GETS) Modelling}

\description{Generic function that enables new GETS methods for new classes}

\usage{
gets(x, ...)

## Default S3 method:
gets.default(x, ...)
}

\arguments{
  \item{x}{an object used to select a method}
  \item{...}{further arguments passed to or from other methods}
}

\author{Genaro Sucarrat, \url{http://www.sucarrat.net/}}

\keyword{Statistical Models}
\keyword{Time Series}
\keyword{Econometrics}
\keyword{Financial Econometrics}

